feature/issue-108-increase-test-coverage-child Review changes
